+++ micrometer-core/src/test/java/io/micrometer/core/instrument/distribution/TimeWindowFixedBoundaryHistogramTest.java
@@ -24,14 +24,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 class TimeWindowFixedBoundaryHistogramTest {
     @Test
     void binarySearchForTail() {
-        assertTailSearch(3, 1, 1L, 5L, 10L);
-        assertTailSearch(5, 1, 1L, 5L, 10L);
-        assertTailSearch(3, 1, 1L, 4L, 5L, 10L);
-        assertTailSearch(3, 2, 1L, 2L, 5L, 10L);
-        assertTailSearch(11, -1, 1L, 5L, 10L);
+        assertTailSearch(3, 1, 1.0, 5.0, 10.0);
+        assertTailSearch(5, 1, 1.0, 5.0, 10.0);
+        assertTailSearch(3, 1, 1.0, 4.0, 5L, 10.0);
+        assertTailSearch(3, 2, 1.0, 2.0, 5L, 10.0);
+        assertTailSearch(11, -1, 1.0, 5.0, 10.0);
     }
 
-    private void assertTailSearch(int search, int expectedIndex, long... buckets) {
+    private void assertTailSearch(int search, int expectedIndex, double... buckets) {
         DistributionStatisticConfig statisticConfig = DistributionStatisticConfig.builder().sla(buckets).build();
         try (TimeWindowFixedBoundaryHistogram histogram = new TimeWindowFixedBoundaryHistogram(Clock.SYSTEM,
                 statisticConfig.merge(DistributionStatisticConfig.DEFAULT), false)) {
@@ -44,7 +44,7 @@ class TimeWindowFixedBoundaryHistogramTest {
     void histogramsAreCumulative() {
         try (TimeWindowFixedBoundaryHistogram histogram = new TimeWindowFixedBoundaryHistogram(new MockClock(),
                 DistributionStatisticConfig.builder()
-                        .sla(3, 6, 7)
+                        .sla(3.0, 6, 7)
                         .bufferLength(1)
                         .build()
                         .merge(DistributionStatisticConfig.DEFAULT), false)) {
@@ -52,9 +52,9 @@ class TimeWindowFixedBoundaryHistogramTest {
             histogram.recordDouble(3);
     
             assertThat(histogram.takeSnapshot(0, 0, 0).histogramCounts()).containsExactly(
-                    new CountAtBucket(3, 1),
-                    new CountAtBucket(6, 1),
-                    new CountAtBucket(7, 1));
+                    new CountAtBucket(3.0, 1),
+                    new CountAtBucket(6.0, 1),
+                    new CountAtBucket(7.0, 1));
     
             histogram.recordDouble(6);
     
@@ -63,9 +63,9 @@ class TimeWindowFixedBoundaryHistogramTest {
             // Proves that the accumulated histogram is truly cumulative, and not just a representation
             // of the last snapshot
             assertThat(histogram.takeSnapshot(0, 0, 0).histogramCounts()).containsExactly(
-                    new CountAtBucket(3, 1),
-                    new CountAtBucket(6, 2),
-                    new CountAtBucket(7, 3)
+                    new CountAtBucket(3.0, 1),
+                    new CountAtBucket(6.0, 2),
+                    new CountAtBucket(7.0, 3)
             );
         }
     }
